Rename and polish IndexWeaviate component
