+++ .pre-commit-config.yaml
@@ -59,11 +59,11 @@ repos:
           - types-retry
         pass_filenames: false
 
-  # - repo: local
-  #   hooks:
-  #     - id: generate_component_readmes
-  #       name: Generate component READMEs
-  #       language: python
-  #       entry: python scripts/component_readme/generate_readme.py
-  #       files: ^components/.*/fondant_component.yaml
-  #       additional_dependencies: ["fondant[aws,azure,gcp]@git+https://github.com/ml6team/fondant@main"]
+  - repo: local
+    hooks:
+      - id: generate_component_readmes
+        name: Generate component READMEs
+        language: python
+        entry: python scripts/component_readme/generate_readme.py
+        files: ^components/.*/fondant_component.yaml
+        additional_dependencies: ["fondant@git+https://github.com/ml6team/fondant@main"]
