+++ components/index_weaviate/Dockerfile
