+++ components/index_weaviate/requirements.txt
